+++ shared/src/usecases/git/generateBranchName.ts
